v0.3.0 Almost ready to deploy
