+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "dongraphio"
-version = "0.2.5"
+version = "0.3.0"
 description = "Small utility library containing graph algorighms used in other projects"
 authors = ["Danila <63115678+DDonnyy@users.noreply.github.com>"]
 readme = "README.md"
@@ -14,7 +14,7 @@ osmnx = "^1.9.1"
 tqdm = "^4.66.2"
 osm2geojson = "^0.2.4"
 pydantic = "^2.6.1"
-momepy = "^0.7.0"
+momepy = "^0.6.0"
 networkit = "^11.0"
 numpy = "^1.23.5"
 pandas = "^2.2.0"
