+++ src/dongraphio/__init__.py
@@ -1,5 +1,12 @@
-__version__ = "0.2.5"
+__version__ = "0.3.0"
 
 from .dongraphio import DonGraphio
 from .enums import GraphType
-from .utils import graphs
+from .utils import (
+    add_projected_points_as_nodes,
+    clusterize_kmeans_geo_points,
+    nx_to_gdf,
+    project_points_on_graph,
+    resolve_tsp,
+    subgraph_by_path,
+)
