+++ src/dongraphio/utils/__init__.py
@@ -1,1 +1,16 @@
-from .graphs import *
+from .geometry_utils import clusterize_kmeans_geo_points
+from .graph_utils import (
+    add_projected_points_as_nodes,
+    convert_multidigraph_to_digraph,
+    get_nearest_edge_geometry,
+    nx_to_gdf,
+    parse_overpass_route_response,
+    project_platforms,
+    project_point_on_edge,
+    project_points_on_graph,
+    subgraph_by_path,
+    update_edges,
+)
+from .matrix_utils import get_dist_matrix, get_dist_matrix_for_tsp, get_nx2nk_idmap, get_subgraph
+from .osm_worker import get_boundary, get_routes, overpass_request
+from .tsp_solver import resolve_tsp
